move weighted scores in api
